+++ packages/react-native-builder-bob/src/build.ts
@@ -1,15 +1,11 @@
-import path from 'path';
-import kleur from 'kleur';
-import * as logger from './utils/logger';
-import buildCommonJS from './targets/commonjs';
-import buildModule from './targets/module';
-import buildTypescript from './targets/typescript';
-import buildCodegen from './targets/codegen';
-import customTarget from './targets/custom';
-import { type Options, type Target } from './types';
 import fs from 'fs-extra';
-import { loadConfig } from './utils/loadConfig';
+import kleur from 'kleur';
+import path from 'path';
 import yargs from 'yargs';
+import { type Options, type Target } from './types';
+import { loadConfig } from './utils/loadConfig';
+import * as logger from './utils/logger';
+import { run } from './utils/workerize';
 
 export const args = {
   target: {
@@ -35,49 +31,45 @@ export async function build(argv: Argv) {
     );
   }
 
-  const result = await loadConfig();
+  const result = loadConfig(root);
 
   if (!result?.config) {
-    logger.error(
+    throw new Error(
       `No configuration found. Run '${argv.$0} init' to create one automatically.`
     );
-    process.exit(1);
   }
 
   const options: Options = result!.config;
 
   if (!options.targets?.length) {
-    logger.error(
-      `No targets found in the configuration in '${path.relative(
+    throw new Error(
+      `No 'targets' found in the configuration in '${path.relative(
         root,
         result!.filepath
       )}'.`
     );
-    process.exit(1);
   }
 
   const source = options.source;
 
   if (!source) {
-    logger.error(
-      `No source option found in the configuration in '${path.relative(
+    throw new Error(
+      `No 'source' option found in the configuration in '${path.relative(
         root,
         result!.filepath
       )}'.`
     );
-    process.exit(1);
   }
 
   const output = options.output;
 
   if (!output) {
-    logger.error(
-      `No source option found in the configuration in '${path.relative(
+    throw new Error(
+      `No 'output' option found in the configuration in '${path.relative(
         root,
         result!.filepath
       )}'.`
     );
-    process.exit(1);
   }
 
   const exclude = options.exclude ?? '**/{__tests__,__fixtures__,__mocks__}/**';
@@ -147,21 +139,11 @@ async function buildTarget({
 
   switch (targetName) {
     case 'commonjs':
-      await buildCommonJS({
-        root,
-        source: path.resolve(root, source),
-        output: path.resolve(root, output, 'commonjs'),
-        exclude,
-        options: targetOptions,
-        variants,
-        report,
-      });
-      break;
     case 'module':
-      await buildModule({
+      await run(targetName, {
         root,
         source: path.resolve(root, source),
-        output: path.resolve(root, output, 'module'),
+        output: path.resolve(root, output, targetName),
         exclude,
         options: targetOptions,
         variants,
@@ -183,7 +165,7 @@ async function buildTarget({
             return false;
           }) ?? false;
 
-        await buildTypescript({
+        await run('typescript', {
           root,
           source: path.resolve(root, source),
           output: path.resolve(root, output, 'typescript'),
@@ -195,7 +177,7 @@ async function buildTarget({
       }
       break;
     case 'codegen':
-      await buildCodegen({
+      await run('codegen', {
         root,
         source: path.resolve(root, source),
         output: path.resolve(root, output, 'typescript'),
@@ -203,7 +185,7 @@ async function buildTarget({
       });
       break;
     case 'custom':
-      await customTarget({
+      await run('custom', {
         options: targetOptions,
         source: path.resolve(root, source),
         report,
@@ -211,7 +193,6 @@ async function buildTarget({
       });
       break;
     default:
-      logger.error(`Invalid target ${kleur.blue(targetName)}.`);
-      process.exit(1);
+      throw new Error(`Invalid target ${kleur.blue(targetName)}.`);
   }
 }